Rename and reformat disasm files
