+++ src/disasm_jumpcall.h
@@ -1,8 +1,9 @@
+
 /*
-	avrdisas - A disassembler for AVR microcontroller units
-	Copyright (C) 2007 Johannes Bauer
-	
-	This file is part of avrdisas.
+    avrdisas - A disassembler for AVR microcontroller units
+    Copyright (C) 2007 Johannes Bauer
+
+    This file is part of avrdisas.
 
     avrdisas is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -18,10 +19,17 @@
     along with avrdisas; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-	Johannes Bauer
-	Mussinanstr. 140
-	92318 Neumarkt/Opf.
-	JohannesBauer@gmx.de
+    Johannes Bauer
+    Mussinanstr. 140
+    92318 Neumarkt/Opf.
+    JohannesBauer@gmx.de
 */
 
-unsigned char BitCount(unsigned char Number);
+void Display_JumpCalls();
+void Register_JumpCall(int From, int To, int Type, unsigned char FunctionCall);
+int JC_Comparison(const void *Element1, const void *Element2);
+void Sort_JumpCalls();
+void Enumerate_Labels();
+char *Get_Label_Name(int Destination, char **LabelComment);
+void Print_JumpCalls(int Position);
+int FixTargetAddress(int Address);
