+++ src/disasm_mnemonics.h
@@ -1,8 +1,9 @@
+
 /*
-	avrdisas - A disassembler for AVR microcontroller units
-	Copyright (C) 2007 Johannes Bauer
-	
-	This file is part of avrdisas.
+    avrdisas - A disassembler for AVR microcontroller units
+    Copyright (C) 2007 Johannes Bauer
+
+    This file is part of avrdisas.
 
     avrdisas is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -18,13 +19,11 @@
     along with avrdisas; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-	Johannes Bauer
-	Mussinanstr. 140
-	92318 Neumarkt/Opf.
-	JohannesBauer@gmx.de
+    Johannes Bauer
+    Mussinanstr. 140
+    92318 Neumarkt/Opf.
+    JohannesBauer@gmx.de
 */
 
 extern const char *MNemonic[];
 extern const char *Cycles[];
-
-
