+++ src/disasm_options.h
@@ -1,8 +1,9 @@
+
 /*
-	avrdisas - A disassembler for AVR microcontroller units
-	Copyright (C) 2007 Johannes Bauer
-	
-	This file is part of avrdisas.
+    avrdisas - A disassembler for AVR microcontroller units
+    Copyright (C) 2007 Johannes Bauer
+
+    This file is part of avrdisas.
 
     avrdisas is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -18,22 +19,11 @@
     along with avrdisas; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-	Johannes Bauer
-	Mussinanstr. 140
-	92318 Neumarkt/Opf.
-	JohannesBauer@gmx.de
+    Johannes Bauer
+    Mussinanstr. 140
+    92318 Neumarkt/Opf.
+    JohannesBauer@gmx.de
 */
 
-#define TYPE_BYTE		1
-#define TYPE_WORD		2
-#define TYPE_ASTRING	3
-#define TYPE_STRING		4
-
-int Read_Tagfile(const char *Filename);
-int Tagfile_FindLabelAddress(int Address);
-char *Tagfile_GetLabel(int TagIndex);
-char *Tagfile_GetLabelComment(int TagIndex);
-int Tagfile_FindPGMAddress(int Address);
-const char* Tagfile_Resolve_Mem_Address(int Address);
-int Tagfile_Process_Data(char *Bitstream, int Position);
-
+void Options_Default(struct Options *Options);
+char Options_ParseCmdLine(struct Options *Options, int argc, char **argv);
