+++ src/disasm_tagfile.h
@@ -1,8 +1,9 @@
+
 /*
-	avrdisas - A disassembler for AVR microcontroller units
-	Copyright (C) 2007 Johannes Bauer
-	
-	This file is part of avrdisas.
+    avrdisas - A disassembler for AVR microcontroller units
+    Copyright (C) 2007 Johannes Bauer
+
+    This file is part of avrdisas.
 
     avrdisas is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -18,18 +19,21 @@
     along with avrdisas; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-	Johannes Bauer
-	Mussinanstr. 140
-	92318 Neumarkt/Opf.
-	JohannesBauer@gmx.de
+    Johannes Bauer
+    Mussinanstr. 140
+    92318 Neumarkt/Opf.
+    JohannesBauer@gmx.de
 */
 
-#ifdef DOS_BASED_OS
-	#define snprintf _snprintf
-#endif
+#define TYPE_BYTE       1
+#define TYPE_WORD       2
+#define TYPE_ASTRING    3
+#define TYPE_STRING     4
 
-#ifdef DESPERATE_SPRINTF_FALLBACK
-	#warning "Your C library does not support snprintf. Please upgrade! Fallback to sprintf used."
-	#define snprintf(dest, len, str...)	sprintf(dest, str)
-#endif
-
+int Read_Tagfile(const char *Filename);
+int Tagfile_FindLabelAddress(int Address);
+char *Tagfile_GetLabel(int TagIndex);
+char *Tagfile_GetLabelComment(int TagIndex);
+int Tagfile_FindPGMAddress(int Address);
+const char *Tagfile_Resolve_Mem_Address(int Address);
+int Tagfile_Process_Data(char *Bitstream, int Position);
